*** VERSION 1.0 RELEASED*** The Jem text editor is finally released! This release comes with some major changes. These changes are...
-Syntax coloring added. This includes special colors for strings, comments,
keywords (C only), search results, and digits. These special colors have
define macros for ease of access.

-Fixed a bug where having a file with unprintable characters crashed the
text editor if one tried scrolling.

-Changed the quit option from cntrl-q to cntrl-x. This was done due to
outside pressures and safe porting (some terminals usually have cntrl-q
as their terminal exit key).

-Changed the way pg up and pg down worked. Previously, one press of either
would send the cursor either at the top of your current screen or at the bottom
of your screen. Still does that in this version, however now if you press
any of those buttons *twice* you scroll just about an entire screen's
worth of lines. If your cursor is already at the top or bottom of your
screen you only need to press the buttons once.

With that, this project comes to an end. While perhaps in the far future
I may change some small things, or perhaps even find a bug or two, as of
right now this project is officially finished. Use this code base if
you really want to, or simply grab it and mess around with it if
you're interest in making a text editor that fits you like a glove.

Signed-off-by: Jose Maldonado <emanueljose11@hotmail.com>